Complete
Everything worked better on this tutorial - except the menu bar is
missing just like on the previous tutorial as well. Trying to still
find out why on both (maybe something to do with most recent update)

The menu bar when directly imported into other projects works fine and
the code between projects is comparably the same. It’s being called and
inflated, but not displaying at the top.